# БР1.2/Москалец Данила/homeworks/hw2/openapi.yaml
openapi: 3.0.0
info:
  title: Sangel API
  description: Маркетплейс охранных услуг
  version: 1.0.0
  contact:
    name: Москалец Данила Алексеевич
servers:
  - url: https://api.sangel.ru/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, USER, OWNER]
        first_name:
          type: string
        last_name:
          type: string
        middle_name:
          type: string
        is_verified:
          type: boolean
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        first_name:
          type: string
        last_name:
          type: string
        middle_name:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            expires_in:
              type: integer
...
    Category:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        is_published:
          type: boolean
    CreateCategoryRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        is_published:
          type: boolean
          default: true
    UpdateCategoryRequest:
      type: object
      properties:
        title:
          type: string
        is_published:
          type: boolean

    # Service
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Регистрация
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: Успешная регистрация
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        400:
          description: Невалидные данные
        409:
          description: Email уже существует
  /auth/login:
    post:
      summary: Вход
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        401:
          description: Неверный email или пароль
  /auth/logout:
    post:
      summary: Выход из системы
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        200:
          description: Успешный выход
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Successfully logged out
        400:
          description: Токен не предоставлен
        401:
          description: Невалидный токен
  /auth/me:
    get:
      summary: Получить текущего пользователя
      tags: [Auth]
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        401:
          description: Не авторизован
  /auth/refresh:
    post:
      summary: Обновить JWT
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        200:
          description: Новый токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        401:
          description: Невалидный refresh token

  # ==================== PROFILE ====================
...
  /categories:
    get:
      summary: Список категорий
      tags: [Categories]
      security: []
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'

    post:
      summary: Создать категорию (только ADMIN)
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        201:
          description: Категория создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
        403:
          description: Доступ запрещен (только ADMIN)
        400:
          description: Невалидные данные
  /categories/{id}:
    put:
      summary: Обновить категорию (только ADMIN)
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        200:
          description: Категория обновлена
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
        403:
          description: Доступ запрещен (только ADMIN)
        404:
          description: Категория не найдена

    delete:
      summary: Удалить категорию (только ADMIN)
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Категория удалена
        403:
          description: Доступ запрещен (только ADMIN)
        404:
          description: Категория не найдена

  # ==================== SERVICES ====================
...